Add new Flexbuf routines: flexready, flexinsstr.
The flexready routine is a crop of the checks originally done in
flexaddch to insure the Flexbuf is ready for manipulation. Those
checks need to be done also for flexinsstr, so make it available
as a separate function.
The flexinsstr bases much of its code on the editable commandline
additions for inserting into the middle of the Flexbuf. This
operation can be used in multiple places, so create a separate
function.